+++ CHANGELOG.rst
@@ -6,8 +6,9 @@ HEAD
 
 **To be included in next release**
 
-  - add table view module
+  - add tableof view module
   - add conversion functions of models to dataframe
+  - add two modes of tableof (resources and models)
   - fix minor bug in bounds collector (unknown collector type)
   - fix templates for generating units